Factored mask_positions implementation and kernels into separate files Doing this will make implementing more accumulators convenient
